moved color blending methods from ExploreAccordionSections.jsx to colorUtils.js
changed font size for Explore Page Accordion Headers and aligned the count number to the top of the line of the title.
To align the count to the top, I separated the number and the accordion Icon (see differences in ExploreSourceGroup).

made explore source icon's position to be always on bottom right (since we separated the count and icon, the icon was floating and it looked bad).

Update explore.css

Update explore.css

Changed the title right color to match the mock-up color. Removed cards and reduced the margin between the explore sources.

moved color blending methods from ExploreAccordionSections.jsx to colorUtils.js

changed font size for Explore Page Accordion Headers and aligned the count number to the top of the line of the title.
To align the count to the top, I separated the number and the accordion Icon (see differences in ExploreSourceGroup).

made explore source icon's position to be always on bottom right (since we separated the count and icon, the icon was floating and it looked bad).

Update explore.css

Update explore.css

Changed the title right color to match the mock-up color. Removed cards and reduced the margin between the explore sources.

Fixed the position of the chevron.